tests: index: extract test_checkout
